Algorithms

Low-Storage Runge–Kutta (LSRK) methods

Low-storage Runger–Kutta methods reduce the number stages that need to be stored. The methods below require only one additional storage vector.

An IncrementingODEProblem must be used.

ClimaTimeSteppers.LowStorageRungeKutta2N — Type

LowStorageRungeKutta2N <: DistributedODEAlgorithm

A class of low-storage Runge-Kutta algorithms, which use only one additional copy of the state vector $u$ (often referred to as $2N$ schemes).

The available implementations are:

LSRKEulerMethod

LSRK54CarpenterKennedy

LSRK144NiegemannDiehlBusch

ClimaTimeSteppers.LSRK54CarpenterKennedy — Type

LSRK54CarpenterKennedy()

The 4th-order, 5-stage [LowStorageRungeKutta2N])(ref) scheme from Solution 3 of Mark H Carpenter, Christopher A Kennedy (1994).

ClimaTimeSteppers.LSRK144NiegemannDiehlBusch — Type

LSRK144NiegemannDiehlBusch()

The 4th-order, 14-stage, [LowStorageRungeKutta2N])(ref) scheme of Jens Niegemann, Richard Diehl, Kurt Busch (2012) with optimized stability region

ClimaTimeSteppers.LSRKEulerMethod — Type

LSRKEulerMethod()

An implementation of explicit Euler method using LowStorageRungeKutta2N infrastructure. This is mainly for debugging.

Strong Stability-Preserving Runge–Kutta (SSPRK) methods

ClimaTimeSteppers.StrongStabilityPreservingRungeKutta — Type

StrongStabilityPreservingRungeKutta <: DistributedODEAlgorithm

A class of Strong Stability Preserving Runge–Kutta methods. These require two additional copies of the state vector.

The available implementations are:

SSPRK22Heuns

SSPRK22Ralstons

SSPRK33ShuOsher

SSPRK34SpiteriRuuth

ClimaTimeSteppers.SSPRK22Heuns — Type

SSPRK22Heuns()

The second-order, 2-stage, strong-stability-preserving, Runge–Kutta scheme of Chi-Wang Shu, Stanley Osher (1988), also known as Heun's method (Heun1900.

Exact choice of coefficients from wikipedia page for Heun's method :)

ClimaTimeSteppers.SSPRK22Ralstons — Type

SSPRK22Ralstons()

The second-order, 2-stage, strong-stability-preserving, Runge–Kutta scheme of Chi-Wang Shu, Stanley Osher (1988), also known as Ralstons's method (Anthony Ralston (1962).

Exact choice of coefficients from wikipedia page for Heun's method :)

ClimaTimeSteppers.SSPRK33ShuOsher — Type

SSPRK33ShuOsher()

The third-order, 3-stage, strong-stability-preserving, Runge–Kutta scheme of Chi-Wang Shu, Stanley Osher (1988).

ClimaTimeSteppers.SSPRK34SpiteriRuuth — Type

SSPRK34SpiteriRuuth()

The third-order, 4-stage, strong-stability-preserving, Runge–Kutta scheme of Raymond J Spiteri, Steven J Ruuth (2002).

Additive Runge–Kutta (ARK) methods

ARK methods are IMEX (Implicit-Explicit) methods based on splitting the ODE function into a linear and remainder components:

\[\frac{du}{dt} = Lu + f_R(u,t)\]

where the linear part is solved implicitly. All the algorithms below take a linsolve argument to specify the linear solver to be used. See the linsolve specification of DifferentialEquations.jl.

Currently ARK methods require a SplitODEProblem.

ClimaTimeSteppers.AdditiveRungeKutta — Type

AdditiveRungeKutta

IMEX (IMplicit-EXplicit) algorithms using ARK (Additively-partitioned Runge-Kutta) methods.

The available implementations are:

ARK1ForwardBackwardEuler

ARK2ImplicitExplicitMidpoint

ARK2GiraldoKellyConstantinescu

ARK437L2SA1KennedyCarpenter

ARK548L2SA2KennedyCarpenter

ClimaTimeSteppers.ARK1ForwardBackwardEuler — Type

ARK1ForwardBackwardEuler(linsolve)

A first-order-accurate two-stage additive Runge–Kutta scheme of Uri M. Ascher, Steven J. Ruuth, Raymond J. Spiteri (1997), combining a forward Euler explicit step with a backward Euler implicit correction.

ClimaTimeSteppers.ARK2ImplicitExplicitMidpoint — Type

ARK2ImplicitExplicitMidpoint(linsolve)

A second-order, two-stage additive Runge–Kutta scheme Uri M. Ascher, Steven J. Ruuth, Raymond J. Spiteri (1997), combining the implicit and explicit midpoint methods.

ClimaTimeSteppers.ARK2GiraldoKellyConstantinescu — Type

ARK2GiraldoKellyConstantinescu(linsolve; paperversion=false)

The second-order, 3-stage additive Runge–Kutta scheme of GKC2013.

If the keyword paperversion=true is used, the coefficients from the paper are used. Otherwise it uses coefficients that make the scheme (much) more stable but less accurate

ClimaTimeSteppers.ARK437L2SA1KennedyCarpenter — Type

ARK437L2SA1KennedyCarpenter(linsolve)

The fourth-order, 7-stage additive Runge–Kutta scheme ARK4(3)7L[2]SA₁ of Christopher A Kennedy, Mark H Carpenter (2019).

ClimaTimeSteppers.ARK548L2SA2KennedyCarpenter — Type

ARK548L2SA2KennedyCarpenter(linsolve)

The fifth-order, 8-stage additive Runge–Kutta scheme ARK5(4)8L[2]SA₂ of Christopher A Kennedy, Mark H Carpenter (2019).

ARS

Missing docstring.

Missing docstring for ARS111. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ARS121. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ARS343. Check Documenter's build log for details.

Multirate

ClimaTimeSteppers.Multirate — Type

Multirate(fast, slow)

A multirate Runge–Kutta scheme, combining fast and slow algorithms

slow can be any algorithm providing methods for the following functions

init_inner(prob, outercache, dt)

update_inner!(innerinteg, outercache, f_slow, u, p, t, dt, stage)

Algorithms which currently support this are:

LowStorageRungeKutta2N

MultirateInfinitesimalStep

WickerSkamarockRungeKutta

Multirate Infinitesimal Step

ClimaTimeSteppers.MultirateInfinitesimalStep — Type

MultirateInfinitesimalStep

Multirate Infinitesimal Step (MIS) methods of J{\"o}rg Wensch, Oswald Knoth, Alexander Galant (2009) and Oswald Knoth, Joerg Wensch (2014).

The available implementations are:

MIS2

MIS3C

MIS4

MIS4a

TVDMISA

TVDMISB

ClimaTimeSteppers.MIS2 — Type

MIS2()

The MIS2 Multirate Infinitesimal Step (MIS) method of Oswald Knoth, Joerg Wensch (2014).

ClimaTimeSteppers.MIS3C — Type

MIS3C()

The MIS3C Multirate Infinitesimal Step (MIS) method of Oswald Knoth, Joerg Wensch (2014).

ClimaTimeSteppers.MIS4 — Type

MIS4()

The MIS4 Multirate Infinitesimal Step (MIS) method of Oswald Knoth, Joerg Wensch (2014).

ClimaTimeSteppers.MIS4a — Type

MIS4a()

The MIS4a Multirate Infinitesimal Step (MIS) method of Oswald Knoth, Joerg Wensch (2014).

ClimaTimeSteppers.TVDMISA — Type

TVDMISA()

The TVDMISA Total Variation Diminishing (TVD) Multirate Infinitesimal Step (MIS) method of Oswald Knoth, Joerg Wensch (2014).

ClimaTimeSteppers.TVDMISB — Type

TVDMISB()

The TVDMISB Total Variation Diminishing (TVD) Multirate Infinitesimal Step (MIS) method of Oswald Knoth, Joerg Wensch (2014).

Wicker–Skamarock

ClimaTimeSteppers.WickerSkamarockRungeKutta — Type

WickerSkamarockRungeKutta <: DistributedODEAlgorithm

Class of multirate algorithms developed in Louis J Wicker, William C Skamarock (1998) and Louis J Wicker, William C Skamarock (2002), which can be used as slow methods in Multirate.

These require two additional copies of the state vector $u$.

Available implementations are:

WSRK2

WSRK3

ClimaTimeSteppers.WSRK2 — Type

WSRK2()

The 2 stage, 2nd order RK2 scheme of Louis J Wicker, William C Skamarock (1998).

ClimaTimeSteppers.WSRK3 — Type

WSRK3()

The 3 stage, 2nd order (3rd order for linear problems) RK3 scheme of Louis J Wicker, William C Skamarock (2002).