Callbacks

ClimaTimeSteppers.Callbacks — Module

ClimaTimeSteppers.Callbacks

A suite of callback functions to be used with the ClimaTimeSteppers.jl ODE solvers.

Interfaces

ClimaTimeSteppers.Callbacks.initialize! — Function

ClimaTimeSteppers.Callbacks.initialize!(f!::F, integrator)

Initialize a callback event for callbacks of type F. By default this does nothing, but can be extended for new callback events.

ClimaTimeSteppers.Callbacks.finalize! — Function

ClimaTimeSteppers.Callbacks.finalize!(f!::F, integrator)

Finalize a callback event for callbacks of type F. By default this does nothing, but can be extended for new callback events.

Callbacks

ClimaTimeSteppers.Callbacks.EveryXWallTimeSeconds — Function

EveryXWallTimeSeconds(
    f!,
    Δwt,
    comm_ctx::ClimaComms.AbstractCommsContext;
    atinit=false
)

Trigger f!(integrator) every Δwt wallclock seconds.

An ClimaComms context must be provided to synchronize timing across all ranks.

Callbacks.initialize! and Callbacks.finalize! can be defined for f!.

If atinit=true, then f!(integrator) will additionally be triggered at initialization, otherwise the first trigger will be after Δwt seconds.

ClimaTimeSteppers.Callbacks.EveryXSimulationTime — Function

EveryXSimulationTime(f!, Δt; atinit=false)

Trigger f!(integrator) every Δt simulation time.

Callbacks.initialize! and Callbacks.finalize! can be defined for f!.

If atinit=true, then f! will additionally be triggered at initialization. Otherwise the first trigger will be after Δt simulation time.

ClimaTimeSteppers.Callbacks.EveryXSimulationSteps — Function

EveryXSimulationSteps(f!, Δsteps; atinit=false)

Trigger f!(integrator) every Δsteps simulation steps.

Callbacks.initialize! and Callbacks.finalize! can be defined for f!.

If atinit==true, then f! will additionally be triggered at initialization. Otherwise the first trigger will be after Δsteps.