Callbacks

ClimaTimeSteppers.Callbacks — Module

ClimaTimeSteppers.Callbacks

A suite of callback functions to be used with the ClimaTimeSteppers.jl ODE solvers.

Interfaces

ClimaTimeSteppers.Callbacks.initialize! — Function

ClimaTimeSteppers.Callbacks.initialize!(f!::F, integrator)

Initialize a callback event for callbacks of type F. By default this does nothing, but can be extended for new callback events.

ClimaTimeSteppers.Callbacks.finalize! — Function

ClimaTimeSteppers.Callbacks.finalize!(f!::F, integrator)

Finalize a callback event for callbacks of type F. By default this does nothing, but can be extended for new callback events.

Callbacks

ClimaTimeSteppers.Callbacks.EveryXWallTimeSeconds — Function

EveryXWallTimeSeconds(f!, Δwt, comm_ctx::ClimaComms.AbstractCommsContext;
                      atinit=false)

Trigger f!(integrator) every Δwt wallclock seconds.

An ClimaComms context must be provided to synchronize timing across all ranks.

Callbacks.initialize! and Callbacks.finalize! can be defined for f!.

If atinit=true, then f!(integrator) will additionally be triggered at initialization, otherwise the first trigger will be after Δwt seconds.

ClimaTimeSteppers.Callbacks.EveryXSimulationTime — Function

EveryXSimulationTime(f!, Δt;
                     atinit=false)

Trigger f!(integrator) every Δt simulation time.

Callbacks.initialize! and Callbacks.finalize! can be defined for f!.

If atinit=true, then f! will additionally be triggered at initialization. Otherwise the first trigger will be after Δt simulation time.

ClimaTimeSteppers.Callbacks.EveryXSimulationSteps — Function

EveryXSimulationSteps(f!, Δsteps;
                     atinit=false)

Trigger f!(integrator) every Δsteps simulation steps.

Callbacks.initialize! and Callbacks.finalize! can be defined for f!.

If atinit==true, then f! will additionally be triggered at initialization. Otherwise the first trigger will be after Δsteps.

Progress Callbacks

ClimaTimeSteppers.BasicProgressCallback — Function

BasicProgressCallback(; kwargs...)

Creates a DiffEqBase.DiscreteCallback that prints out a progress bar as the integrator runs.

Keywords

io::IO = stdout: output stream to which the progress bar is printed

min_update_time::Real = 0.5: minimum delay (in seconds) between prints

bar_title::String = "Progress: ": string that gets printed to the left of the progress bar

bar_length::Integer = 100: number of characters in the progress bar

show_tens_intervals::Bool = true: whether to add ticks above the progress bar that indicate the positions of 0%, 10%, 20%, ..., 100%; if this is false, only the ticks at 0% and 100% are shown

ClimaTimeSteppers.TerminalProgressCallback — Function

TerminalProgressCallback([tType]; kwargs...)

Creates a DiffEqBase.DiscreteCallback that prints out a progress bar as the integrator runs, along with an estimate of the time remaining until the integrator is finished, and also an optional user-specified message (e.g., diagnostic information about the integrator's state, or a Unicode plot).

This progress bar is designed to be overwritten on every update, so it should only be printed to a UNIX terminal (i.e., a terminal that supports clearing the previous line by printing the control sequence "\e[1A\r\e[0K").

The time remaining is estimated by computing an exponential moving average of the integrator's speed (seconds of real time that elapse per unit of integrator time) and multiplying this average speed by the remaining integrator time; the first integrator step is not included in the average so as to avoid biasing the estimate with compilation time.

Arguments

tType::Type = Float64: type of integrator.t

Keywords

io::IO = stdout: terminal output stream to which the progress bar is printed

min_update_time::Real = 0.5: minimum delay (in seconds) between prints

bar_title::String = "Progress: ": string that gets printed to the left of the progress bar

relative_bar_length::Real = 0.7: ratio between the number of characters in the progress bar and the width of the terminal (displaysize(io)[2])

eta_title::String = "Time Remaining: ": string that gets printed to the left of the time remaining

new_speed_weight::Real = 0.1: weight of the new speed in the formula for updating the exponential moving average — average := new_speed_weight * new_speed + (1 - new_speed_weight) * average

custom_message::Union{Nothing, Function} = nothing: an optional function of the form (integrator, terminal_width) -> String that generates a custom message whenever the progress bar is updated; the generated string can have multiple lines, and it is recommended to limit the length of each line to be no bigger than the terminal width in order to improve readability

clear_when_finished::Bool = true: whether to clear away the progress bar (and any other information that was printed) when the integrator is finished