Algorithms

Low-Storage Runge–Kutta (LSRK) methods

Low-storage Runger–Kutta methods reduce the number stages that need to be stored. The methods below require only one additional storage vector.

An IncrementingODEProblem must be used.

TimeMachine.LSRK54CarpenterKennedy — Type

LSRK54CarpenterKennedy()

The fourth-order, 5-stage low storage Runge–Kutta scheme from Solution 3 of Mark H Carpenter , Christopher A Kennedy (1994).

TimeMachine.LSRK144NiegemannDiehlBusch — Type

LSRK144NiegemannDiehlBusch()

The fourth-order, 14-stage, low-storage, Runge–Kutta scheme of Jens Niegemann , Richard Diehl , Kurt Busch (2012) with optimized stability region

TimeMachine.LSRKEulerMethod — Type

LSRKEulerMethod()

An implementation of explicit Euler method using LowStorageRungeKutta2N infrastructure. This is mainly for debugging.

Strong Stability-Preserving Runge–Kutta (SSPRK) methods

TimeMachine.SSPRK22Heuns — Type

SSPRK22Heuns()

The second-order, 2-stage, strong-stability-preserving, Runge–Kutta scheme of Chi-Wang Shu , Stanley Osher (1988), also known as Heun's method (Heun1900.

Exact choice of coefficients from wikipedia page for Heun's method :)

TimeMachine.SSPRK22Ralstons — Type

SSPRK22Ralstons()

The second-order, 2-stage, strong-stability-preserving, Runge–Kutta scheme of Chi-Wang Shu , Stanley Osher (1988), also known as Ralstons's method (Anthony Ralston (1962).

Exact choice of coefficients from wikipedia page for Heun's method :)

TimeMachine.SSPRK33ShuOsher — Type

SSPRK33ShuOsher()

The third-order, 3-stage, strong-stability-preserving, Runge–Kutta scheme of Chi-Wang Shu , Stanley Osher (1988).

TimeMachine.SSPRK34SpiteriRuuth — Type

SSPRK34SpiteriRuuth()

The third-order, 4-stage, strong-stability-preserving, Runge–Kutta scheme of Raymond J Spiteri , Steven J Ruuth (2002).

Additive Runge–Kutta (ARK) methods

ARK methods are IMEX (Implicit-Explicit) methods based on splitting the ODE function into a linear and remainder components:

\[\frac{du}{dt} = Lu + f_R(u,t)\]

where the linear part is solved implicitly. All the algorithms below take a linsolve argument to specify the linear solver to be used. See the linsolve specification of DifferentialEquations.jl.

Currently ARK methods require a SplitODEProblem.

TimeMachine.ARK1ForwardBackwardEuler — Type

ARK1ForwardBackwardEuler(linsolve)

A first-order-accurate two-stage additive Runge–Kutta scheme of Uri M. Ascher , Steven J. Ruuth , Raymond J. Spiteri (1997), combining a forward Euler explicit step with a backward Euler implicit correction.

TimeMachine.ARK2ImplicitExplicitMidpoint — Type

ARK2ImplicitExplicitMidpoint(linsolve)

A second-order, two-stage additive Runge–Kutta scheme Uri M. Ascher , Steven J. Ruuth , Raymond J. Spiteri (1997), combining the implicit and explicit midpoint methods.

TimeMachine.ARK2GiraldoKellyConstantinescu — Type

ARK2GiraldoKellyConstantinescu(linsolve; paperversion=false)

The second-order, 3-stage additive Runge–Kutta scheme of GKC2013.

If the keyword paperversion=true is used, the coefficients from the paper are used. Otherwise it uses coefficients that make the scheme (much) more stable but less accurate

TimeMachine.ARK437L2SA1KennedyCarpenter — Type

ARK437L2SA1KennedyCarpenter(linsolve)

The fourth-order, 7-stage additive Runge–Kutta scheme ARK4(3)7L[2]SA₁ of Christopher A Kennedy , Mark H Carpenter (2019).

TimeMachine.ARK548L2SA2KennedyCarpenter — Type

ARK548L2SA2KennedyCarpenter(linsolve)

The fifth-order, 8-stage additive Runge–Kutta scheme ARK5(4)8L[2]SA₂ of Christopher A Kennedy , Mark H Carpenter (2019).