Algorithms

Low-Storage Runge–Kutta (LSRK) methods

Low-storage Runger–Kutta methods reduce the number stages that need to be stored. The methods below require only one additional storage vector.

An IncrementingODEProblem must be used.

TimeMachine.LSRK54CarpenterKennedy — Type

LSRK54CarpenterKennedy()

The fourth-order, 5-stage low storage Runge–Kutta scheme from Carpenter and Kennedy (1994). The coefficients are those from Solution 3 in the paper.

References

Carpenter, M.H.; Kennedy, C.A. (1994) "Fourth-order $2N$-storage Runge–Kutta schemes", NASA Technical Memorandum 109112. http://hdl.handle.net/2060/19940028444

TimeMachine.LSRK144NiegemannDiehlBusch — Type

LSRK144NiegemannDiehlBusch()

The fourth-order, 14-stage, low-storage, Runge–Kutta scheme of Niegemann, Diehl, and Busch (2012) with optimized stability region

References

Niegemann, J.; Diehl, R.; Busch, K. (2012) "Efficient low-storage Runge–Kutta schemes with optimized stability regions", Journal of Computational Physics 231(2): 364–372. doi: 10.1016/j.jcp.2011.09.003

TimeMachine.LSRKEulerMethod — Type

LSRKEulerMethod()

An implementation of explicit Euler method using LowStorageRungeKutta2N infrastructure. This is mainly for debugging.

Strong Stability-Preserving Runge–Kutta (SSPRK) methods

TimeMachine.SSPRK22Heuns — Type

SSPRK22Heuns()

The second-order, 2-stage, strong-stability-preserving, Runge–Kutta scheme of Shu and Osher (1988) (also known as Heun's method.) Exact choice of coefficients from wikipedia page for Heun's method :)

References

@article{shu1988efficient,
  title={Efficient implementation of essentially non-oscillatory shock-capturing schemes},
  author={Shu, Chi-Wang and Osher, Stanley},
  journal={Journal of computational physics},
  volume={77},
  number={2},
  pages={439--471},
  year={1988},
  publisher={Elsevier}
}
@article {Heun1900,
   title = {Neue Methoden zur approximativen Integration der
   Differentialgleichungen einer unabh"{a}ngigen Ver"{a}nderlichen}
   author = {Heun, Karl},
   journal = {Z. Math. Phys},
   volume = {45},
   pages = {23--38},
   year = {1900}
}

TimeMachine.SSPRK22Ralstons — Type

SSPRK22Ralstons()

The second-order, 2-stage, strong-stability-preserving, Runge–Kutta scheme of Shu and Osher (1988) (also known as Ralstons's method.) Exact choice of coefficients from wikipedia page for Heun's method :)

References

@article{shu1988efficient,
  title={Efficient implementation of essentially non-oscillatory shock-capturing schemes},
  author={Shu, Chi-Wang and Osher, Stanley},
  journal={Journal of computational physics},
  volume={77},
  number={2},
  pages={439--471},
  year={1988},
  publisher={Elsevier}
}
@article{ralston1962runge,
  title={Runge-Kutta methods with minimum error bounds},
  author={Ralston, Anthony},
  journal={Mathematics of computation},
  volume={16},
  number={80},
  pages={431--437},
  year={1962},
  doi={10.1090/S0025-5718-1962-0150954-0}
}

TimeMachine.SSPRK33ShuOsher — Type

SSPRK33ShuOsher()

The third-order, 3-stage, strong-stability-preserving, Runge–Kutta scheme of Shu and Osher (1988)

References

@article{shu1988efficient,
  title={Efficient implementation of essentially non-oscillatory shock-capturing schemes},
  author={Shu, Chi-Wang and Osher, Stanley},
  journal={Journal of computational physics},
  volume={77},
  number={2},
  pages={439--471},
  year={1988},
  publisher={Elsevier}
}

TimeMachine.SSPRK34SpiteriRuuth — Type

SSPRK34SpiteriRuuth()

The third-order, 4-stage, strong-stability-preserving, Runge–Kutta scheme of Spiteri and Ruuth (1988)

References

@article{spiteri2002new,
  title={A new class of optimal high-order strong-stability-preserving time discretization methods},
  author={Spiteri, Raymond J and Ruuth, Steven J},
  journal={SIAM Journal on Numerical Analysis},
  volume={40},
  number={2},
  pages={469--491},
  year={2002},
  publisher={SIAM}
}

Additive Runge–Kutta (ARK) methods

ARK methods are IMEX (Implicit-Explicit) methods based on splitting the ODE function into a linear and remainder components:

\[\frac{du}{dt} = Lu + f_R(u,t)\]

where the linear part is solved implicitly. All the algorithms below take a linsolve argument to specify the linear solver to be used. See the linsolve specification of DifferentialEquations.jl.

Currently ARK methods require a SplitODEProblem.

TimeMachine.ARK1ForwardBackwardEuler — Type

ARK1ForwardBackwardEuler(linsolve)

A first-order-accurate two-stage additive Runge–Kutta scheme, combining a forward Euler explicit step with a backward Euler implicit correction.

References

Ascher, U.M.; Ruuth, S.J. and Spiteri, R.S. (1997) "Implicit-explicit Runge-Kutta methods for time-dependent partial differential equations", Applied Numerical Mathematics, 25(2-3): 151–167. doi: 10.1016/S0168-9274(97)00056-1

TimeMachine.ARK2ImplicitExplicitMidpoint — Type

ARK2ImplicitExplicitMidpoint(linsolve)

A second-order, two-stage additive Runge–Kutta scheme, combining the implicit and explicit midpoint methods.

References

Ascher, U.M.; Ruuth, S.J. and Spiteri, R.S. (1997) "Implicit-explicit Runge-Kutta methods for time-dependent partial differential equations", Applied Numerical Mathematics, 25(2-3): 151–167. doi: 10.1016/S0168-9274(97)00056-1

TimeMachine.ARK2GiraldoKellyConstantinescu — Type

ARK2GiraldoKellyConstantinescu(linsolve; paperversion=false)

The second-order, 3-stage additive Runge–Kutta scheme of Giraldo, Kelly and Constantinescu (2013).

If the keyword paperversion=true is used, the coefficients from the paper are used. Otherwise it uses coefficients that make the scheme (much) more stable but less accurate

References

Giraldo, F.X.; Kelly, J.F. and Constantinescu, E.M. (2013) "Implicit-explicit formulations of a three-dimensional nonhydrostatic unified model of the atmosphere (NUMA)", SIAM Journal on Scientific Computing, 35(5): B1162–B1194 doi: 10.1137/120876034

TimeMachine.ARK437L2SA1KennedyCarpenter — Type

ARK437L2SA1KennedyCarpenter(linsolve)

The fourth-order, 7-stage additive Runge–Kutta scheme ARK4(3)7L[2]SA₁ of Kennedy and Carpenter (2019).

References

Kennedy, C.A. and Carpenter, M.H. (2019) "Higher-order additive Runge–Kutta schemes for ordinary differential equations" Applied Numerical Mathematics. 136: 183–205. doi: 10.1016/j.apnum.2018.10.007

TimeMachine.ARK548L2SA2KennedyCarpenter — Type

ARK548L2SA2KennedyCarpenter(linsolve)

The fifth-order, 8-stage additive Runge–Kutta scheme ARK5(4)8L[2]SA₂ of Kennedy and Carpenter (2019).

References

Kennedy, C.A. and Carpenter, M.H. (2019) "Higher-order additive Runge–Kutta schemes for ordinary differential equations" Applied Numerical Mathematics. 136: 183–205. doi: 10.1016/j.apnum.2018.10.007