ODE Solvers

Tableau Interface

ClimaTimeSteppers.RKTableau — Type

RKTableau(α, a, b, c)

A container for all of the information required to formulate a Runge-Kutta (RK) timestepping method. The arrays a, b, and c comprise the Butcher tableau of the method, while α is either nothing or the first matrix of the method's canonical Shu-Osher formulation (if at least one such formulation is available).

ClimaTimeSteppers.ButcherTableau — Function

ButcherTableau(a, [b], [c])

Constructs an RKTableau without a Shu-Osher formulation, under the default assumptions that it is first-same-as-last (FSAL) and internally consistent.

ClimaTimeSteppers.ShuOsherTableau — Function

ShuOsherTableau(α, a, [b], [c])

Constructs an RKTableau with a canonical Shu-Osher formulation whose first matrix is given by α, under the default assumptions that it is first-same-as-last (FSAL) and internally consistent.

ClimaTimeSteppers.PaddedTableau — Function

PaddedTableau(tableau)

Constructs an RKTableau that is identical to the given RKTableau, but with an additional "empty" stage at the beginning of the method that leaves the initial state unmodified.

ClimaTimeSteppers.ARKTableau — Type

ARKTableau(lim, exp, imp)
ARKTableau(lim_and_exp, imp)
ARKTableau(lim_and_exp)

A container for all of the information required to formulate an additive Runge-Kutta (ARK) timestepping method with three components:

lim, an RKTableau applied to the limited tendency T_lim!

exp, an RKTableau applied to the explicit tendency T_exp!

imp, an RKTableau applied to the implicit tendency T_imp!

If either the lim or exp tableau is not specified, they are assumed to be identical. If the imp tableau is also not specified, it is assumed to be the same as the lim and exp tableaus.

The exp tableau must describe an explicit Runge-Kutta (ERK) method, and the imp tableau must describe either a diagonally implicit Runge-Kutta (DIRK) method or an ERK method. We also require the lim tableau to describe an ERK method, but we could potentially extend this to DIRK methods in the future.

If the lim tableau includes a canonical Shu-Osher formulation matrix, that formulation is used to apply limiters. Otherwise, limiters can only be applied approximately, without any guarantees that monotonicity will be preserved.

ClimaTimeSteppers.is_ERK — Function

is_ERK(tableau)

Checks whether an RKTableau is explicit; i.e., whether its coefficient matrices are strictly lower triangular.

ClimaTimeSteppers.is_DIRK — Function

is_DIRK(tableau)

Checks whether an RKTableau is diagonally implicit; i.e., whether its coefficient matrices are non-strictly lower triangular.

Algorithm Interface

ClimaTimeSteppers.AbstractAlgorithmName — Type

AbstractAlgorithmName

Supertype of predefined Runge-Kutta methods.

ClimaTimeSteppers.RKAlgorithmName — Type

RKAlgorithmName

An AbstractAlgorithmName with a method of the form RKTableau(name).

ClimaTimeSteppers.SSPRKAlgorithmName — Type

SSPRKAlgorithmName

An RKAlgorithmName whose tableau has a canonical Shu-Osher formulation.

ClimaTimeSteppers.ARKAlgorithmName — Type

ARKAlgorithmName

An AbstractAlgorithmName with a method of the form ARKTableau(name).

ClimaTimeSteppers.IMEXSSPRKAlgorithmName — Type

IMEXSSPRKAlgorithmName

An ARKAlgorithmName whose lim tableau has a canonical Shu-Osher formulation.

ClimaTimeSteppers.RKAlgorithm — Function

RKAlgorithm(tableau)
RKAlgorithm(name)

Constructs a Runge-Kutta (RK) algorithm for solving ODEs. The first constructor accepts any RKTableau and leaves the algorithm unnamed, while the second determines the tableau from an RKAlgorithmName. Each of these constructors just makes an ARKAlgorithm with identical lim, exp, and imp tableaus.

ClimaTimeSteppers.ARKAlgorithm — Type

ARKAlgorithm(tableau, [newtons_method])
ARKAlgorithm(name, [newtons_method])

Constructs an additive Runge-Kutta (ARK) algorithm for solving ODEs. The first constructor accepts any ARKTableau and leaves the algorithm unnamed, while the second determines the tableau from an ARKAlgorithmName. If the specified imp tableau necessitates the use of an implicit solver for the problem that this algorithm will solve, a NewtonsMethod must also be specified.

RK Algorithm Names

ClimaTimeSteppers.SSP22Heuns — Type

SSP22Heuns

An SSPRK algorithm from [8], with 2 stages and 2nd order accuracy. Also called Heun's method ([9]).

ClimaTimeSteppers.SSP33ShuOsher — Type

SSP33ShuOsher

An SSPRK algorithm from [8], with 3 stages and 3rd order accuracy.

ClimaTimeSteppers.RK4 — Type

RK4

The RK4 algorithm from [10], a Runge-Kutta method with 4 stages and 4th order accuracy.

ARK Algorithm Names

ClimaTimeSteppers.ARS111 — Type

ARS111

An IMEX ARK algorithm from [11], section 2, with 1 implicit stage, 1 explicit stage and 1st order accuracy. Also called IMEX Euler or forward-backward Euler; equivalent to OrdinaryDiffEq.IMEXEuler.

ClimaTimeSteppers.ARS121 — Type

ARS121

An IMEX ARK algorithm from [11], section 2, with 1 implicit stage, 2 explicit stages, and 1st order accuracy. Also called IMEX Euler or forward-backward Euler; equivalent to OrdinaryDiffEq.IMEXEulerARK.

ClimaTimeSteppers.ARS122 — Type

ARS122

An IMEX ARK algorithm from [11], section 2, with 1 implicit stage, 2 explicit stages, and 2nd order accuracy. Also called IMEX midpoint.

ClimaTimeSteppers.ARS233 — Type

ARS233

An IMEX ARK algorithm from [11], section 2, with 2 implicit stages, 3 explicit stages, and 3rd order accuracy.

ClimaTimeSteppers.ARS232 — Type

ARS232

An IMEX ARK algorithm from [11], section 2, with 2 implicit stages, 3 explicit stages, and 2nd order accuracy.

ClimaTimeSteppers.ARS222 — Type

ARS222

An IMEX ARK algorithm from [11], section 2, with 2 implicit stages, 2 explicit stages, and 2nd order accuracy.

ClimaTimeSteppers.ARS343 — Type

ARS343

An IMEX ARK algorithm from [11], section 2, with 3 implicit stages, 4 explicit stages, and 3rd order accuracy.

ClimaTimeSteppers.ARS443 — Type

ARS443

An IMEX ARK algorithm from [11], section 2, with 4 implicit stages, 4 explicit stages, and 3rd order accuracy.

ClimaTimeSteppers.IMKG232a — Type

IMKG232a

An IMEX ARK algorithm from [12], Table 3, with 2 implicit stages, 3 explicit stages, and 2nd order accuracy.

ClimaTimeSteppers.IMKG232b — Type

IMKG232b

An IMEX ARK algorithm from [12], Table 3, with 2 implicit stages, 3 explicit stages, and 2nd order accuracy.

ClimaTimeSteppers.IMKG242a — Type

IMKG242a

An IMEX ARK algorithm from [12], Table 3, with 2 implicit stages, 4 explicit stages, and 2nd order accuracy.

ClimaTimeSteppers.IMKG242b — Type

IMKG242b

An IMEX ARK algorithm from [12], Table 3, with 2 implicit stages, 4 explicit stages, and 2nd order accuracy.

ClimaTimeSteppers.IMKG243a — Type

IMKG243a

An IMEX ARK algorithm from [12], Table 3, with 3 implicit stages, 4 explicit stages, and 2nd order accuracy.

ClimaTimeSteppers.IMKG252a — Type

IMKG252a

An IMEX ARK algorithm from [12], Table 3, with 2 implicit stages, 5 explicit stages, and 2nd order accuracy.

ClimaTimeSteppers.IMKG252b — Type

IMKG252b

An IMEX ARK algorithm from [12], Table 3, with 2 implicit stages, 5 explicit stages, and 2nd order accuracy.

ClimaTimeSteppers.IMKG253a — Type

IMKG253a

An IMEX ARK algorithm from [12], Table 3, with 3 implicit stages, 5 explicit stages, and 2nd order accuracy.

ClimaTimeSteppers.IMKG253b — Type

IMKG253b

An IMEX ARK algorithm from [12], Table 3, with 3 implicit stages, 5 explicit stages, and 2nd order accuracy.

ClimaTimeSteppers.IMKG254a — Type

IMKG254a

An IMEX ARK algorithm from [12], Table 3, with 4 implicit stages, 5 explicit stages, and 2nd order accuracy.

ClimaTimeSteppers.IMKG254b — Type

IMKG254b

An IMEX ARK algorithm from [12], Table 3, with 4 implicit stages, 5 explicit stages, and 2nd order accuracy.

ClimaTimeSteppers.IMKG254c — Type

IMKG254c

An IMEX ARK algorithm from [12], Table 3, with 4 implicit stages, 5 explicit stages, and 2nd order accuracy.

ClimaTimeSteppers.IMKG342a — Type

IMKG342a

An IMEX ARK algorithm from [12], Table 4, with 2 implicit stages, 4 explicit stages, and 3rd order accuracy.

ClimaTimeSteppers.IMKG343a — Type

IMKG343a

An IMEX ARK algorithm from [12], Table 4, with 3 implicit stages, 4 explicit stages, and 3rd order accuracy.

ClimaTimeSteppers.SSP222 — Type

SSP222

An IMEX SSPRK algorithm from [13], with 2 implicit stages, 2 explicit stages, and 2nd order accuracy. Also called SSP2(222) in [14].

ClimaTimeSteppers.SSP322 — Type

SSP322

An IMEX SSPRK algorithm from [13], with 3 implicit stages, 2 explicit stages, and 2nd order accuracy.

ClimaTimeSteppers.SSP332 — Type

SSP332

An IMEX SSPRK algorithm from [13], with 3 implicit stages, 3 explicit stages, and 2nd order accuracy. Also called SSP2(332)a in [14].

ClimaTimeSteppers.SSP333 — Type

SSP333(; β = 1/2 + √3/6)

Family of IMEX SSPRK algorithms parametrized by the value β from [15], Section 3.2, with 3 implicit stages, 3 explicit stages, and 3rd order accuracy. The default value of β results in an SDIRK algorithm, which is also called SSP3(333)c in [14].

ClimaTimeSteppers.SSP433 — Type

SSP433

An IMEX SSPRK algorithm from [13], with 4 implicit stages, 3 explicit stages, and 3rd order accuracy. Also called SSP3(433) in [14].

ClimaTimeSteppers.DBM453 — Type

DBM453

An IMEX ARK algorithm from [16], Appendix A, with 4 implicit stages, 5 explicit stages, and 3rd order accuracy.

ClimaTimeSteppers.HOMMEM1 — Type

HOMMEM1

An IMEX ARK algorithm from [17], section 4.1, with 5 implicit stages, 6 explicit stages, and 2nd order accuracy.

ClimaTimeSteppers.ARK2GKC — Type

ARK2GKC(; paper_version = false)

An IMEX ARK algorithm from [18] with 2 implicit stages, 3 explicit stages, and 2nd order accuracy. If paper_version = true, the algorithm uses coefficients from the paper. Otherwise, it uses coefficients that make it more stable but less accurate.

ClimaTimeSteppers.ARK437L2SA1 — Type

ARK437L2SA1

An IMEX ARK algorithm from [19], Table 8, with 6 implicit stages, 7 explicit stages, and 4th order accuracy. Written as ARK4(3)7L[2]SA₁ in the paper.

ClimaTimeSteppers.ARK548L2SA2 — Type

ARK548L2SA2

An IMEX ARK algorithm from [19], Table 8, with 7 implicit stages, 8 explicit stages, and 5th order accuracy. Written as ARK5(4)8L[2]SA₂ in the paper.

ClimaTimeSteppers.SSPKnoth — Type

SSPKnoth

SSPKnoth is a third-order Rosenbrock method developed by Oswald Knoth. When integrating an implicit tendency, this reduces to a second-order method because it only performs an approximate implicit solve on each stage.

The coefficients are the same as in CGDycore.jl, except that for C we add the diagonal elements too. Note, however, that the elements on the diagonal of C do not really matter because C is only used in its lower triangular part. We add them mostly to match literature on the subject

Old LSRK Interface

ClimaTimeSteppers.ForwardEulerODEFunction — Type

ForwardEulerODEFunction(f; jac_prototype, Wfact, tgrad)

An ODE function wrapper where f(un, u, p, t, dt) provides a forward Euler update

un .= u .+ dt * f(u, p, t)

ClimaTimeSteppers.LowStorageRungeKutta2N — Type

LowStorageRungeKutta2N <: DistributedODEAlgorithm

A class of low-storage Runge-Kutta algorithms, which use only one additional copy of the state vector $u$ (often referred to as $2N$ schemes).

The available implementations are:

LSRKEulerMethod

LSRK54CarpenterKennedy

LSRK144NiegemannDiehlBusch

ClimaTimeSteppers.LSRK54CarpenterKennedy — Type

LSRK54CarpenterKennedy()

The 4th-order, 5-stage [LowStorageRungeKutta2N])(ref) scheme from Solution 3 of [20].

ClimaTimeSteppers.LSRK144NiegemannDiehlBusch — Type

LSRK144NiegemannDiehlBusch()

The 4th-order, 14-stage, [LowStorageRungeKutta2N])(ref) scheme of [21] with optimized stability region

ClimaTimeSteppers.LSRKEulerMethod — Type

LSRKEulerMethod()

An implementation of explicit Euler method using LowStorageRungeKutta2N infrastructure. This is mainly for debugging.

Old Multirate Interface

ClimaTimeSteppers.Multirate — Type

Multirate(fast, slow)

A multirate Runge–Kutta scheme, combining fast and slow algorithms

slow can be any algorithm providing methods for the following functions

init_inner(prob, outercache, dt)

update_inner!(innerinteg, outercache, f_slow, u, p, t, dt, stage)

Algorithms which currently support this are:

LowStorageRungeKutta2N

MultirateInfinitesimalStep

WickerSkamarockRungeKutta

ClimaTimeSteppers.MultirateInfinitesimalStep — Type

MultirateInfinitesimalStep

Multirate Infinitesimal Step (MIS) methods of [4] and [5].

The available implementations are:

MIS2

MIS3C

MIS4

MIS4a

TVDMISA

TVDMISB

ClimaTimeSteppers.MIS2 — Type

MIS2()

The MIS2 Multirate Infinitesimal Step (MIS) method of [5].

ClimaTimeSteppers.MIS3C — Type

MIS3C()

The MIS3C Multirate Infinitesimal Step (MIS) method of [5].

ClimaTimeSteppers.MIS4 — Type

MIS4()

The MIS4 Multirate Infinitesimal Step (MIS) method of [5].

ClimaTimeSteppers.MIS4a — Type

MIS4a()

The MIS4a Multirate Infinitesimal Step (MIS) method of [5].

ClimaTimeSteppers.TVDMISA — Type

TVDMISA()

The TVDMISA Total Variation Diminishing (TVD) Multirate Infinitesimal Step (MIS) method of [5].

ClimaTimeSteppers.TVDMISB — Type

TVDMISB()

The TVDMISB Total Variation Diminishing (TVD) Multirate Infinitesimal Step (MIS) method of [5].

ClimaTimeSteppers.WickerSkamarockRungeKutta — Type

WickerSkamarockRungeKutta <: DistributedODEAlgorithm

Class of multirate algorithms developed in [6] and [7], which can be used as slow methods in Multirate.

These require two additional copies of the state vector $u$.

Available implementations are:

WSRK2

WSRK3

ClimaTimeSteppers.WSRK2 — Type

WSRK2()

The 2 stage, 2nd order RK2 scheme of [6].

ClimaTimeSteppers.WSRK3 — Type

WSRK3()

The 3 stage, 2nd order (3rd order for linear problems) RK3 scheme of [7].